+++ src/main/java/com/novelbot/api/repository/EpisodeRepository.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EpisodeRepository extends JpaRepository<Episode, Long> {
-    List<Episode> findByNovelNovelId(Long novelId);
+public interface EpisodeRepository extends JpaRepository<Episode, int> {
+    List<Episode> findByNovelNovelId(int novelId);
 }
